pyproject.toml
[project]
name = "status-app-agent"
version = "0.1.0"
description = "Privacy-preserving AI agents for Status Communities"
requires-python = ">=3.12"
dependencies = [
    "websockets>=14.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "anthropic>=0.40",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 120
